# openapi.yaml
openapi: 3.0.0
info:
  title: Second Brain Server API
  description: |
    Comprehensive personal intelligence platform API for knowledge management, creativity, learning, and personal growth.

    ## Features
    - **Stateless Authentication**: JWT-based authentication with no server-side sessions
    - **OAuth Integration**: Google OAuth support with CSRF protection
    - **Database Management**: Create and manage custom databases with flexible schemas
    - **Record Management**: CRUD operations for database records with rich property types
    - **User Management**: Complete user lifecycle management with role-based access control
    - **Rate Limiting**: Built-in rate limiting for security and performance
    - **File Operations**: Import/export data in multiple formats (JSON, CSV, Excel)
    - **Real-time Collaboration**: Share databases with granular permissions

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - Login: 5 attempts per 15 minutes
    - Registration: 3 attempts per hour
    - Password Reset: 3 attempts per hour
    - Token Refresh: 10 attempts per 15 minutes
    - OAuth: 20 attempts per 15 minutes
    - General API: 100 requests per 15 minutes

    ## Database Property Types
    - **TEXT**: Single-line text
    - **NUMBER**: Numeric values
...
    - **DATE**: Date and time values
...
  version: 1.0.0
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: john_doe
        role:
          type: string
          enum: [admin, user, moderator]
          example: user
        isActive:
          type: boolean
          example: true
        authProvider:
          type: string
          enum: [LOCAL, GOOGLE]
          example: LOCAL
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        profilePicture:
          type: string
          format: uri
          example: https://example.com/avatar.jpg
        isEmailVerified:
          type: boolean
          example: true
        lastLoginAt:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: object
          properties:
            message:
              type: string
              example: Error description
            statusCode:
              type: integer
              example: 400
            status:
              type: string
              example: error
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Get all users with filtering and pagination
      description: |
        Retrieve a paginated list of users with advanced filtering options.

        **Features**:
        - Search by email, username, first name, or last name
        - Filter by role, auth provider, and active status
        - Sort by various fields with ascending/descending order
        - Pagination support

        **Required Permission**: Moderator or Admin
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Number of users per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: search
          in: query
          description: Search term for email, username, first name, or last name
          schema:
            type: string
            example: john
        - name: role
          in: query
          description: Filter by user role
          schema:
            type: string
            enum: [USER, MODERATOR, ADMIN]
            example: USER
        - name: authProvider
          in: query
          description: Filter by authentication provider
          schema:
            type: string
            enum: [LOCAL, GOOGLE]
            example: LOCAL
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: boolean
            example: true
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [createdAt, updatedAt, lastLoginAt, email, username]
            default: createdAt
            example: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
            example: desc
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          total:
                            type: integer
                            description: Total number of users matching filters
                            example: 150
                          totalPages:
                            type: integer
                            description: Total number of pages
                            example: 15
                          currentPage:
                            type: integer
                            description: Current page number
                            example: 1
              examples:
                success:
                  summary: Users list with pagination
                  value:
                    success: true
                    message: Users retrieved successfully
                    data:
                      users:
                        - id: 507f1f77bcf86cd799439011
                          email: john.doe@example.com
                          username: john_doe
                          firstName: John
                          lastName: Doe
                          role: USER
                          isActive: true
                          authProvider: LOCAL
                          isEmailVerified: true
                          createdAt: 2023-01-01T00:00:00Z
                          updatedAt: 2023-01-01T12:00:00Z
                          lastLoginAt: 2023-01-01T12:00:00Z
                      total: 150
                      totalPages: 15
                      currentPage: 1
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden (Moderator/Admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /users/stats:
    get:
      tags: [Users]
      summary: Get user statistics
      description: |
        Retrieve comprehensive user statistics for administrative purposes.

        **Features**:
        - Total and active user counts
        - New user registrations in period
        - User distribution by role and auth provider
        - Recent activity trends

        **Required Permission**: Admin only
      operationId: getUserStats
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          description: Time period for statistics
          schema:
            type: string
            enum: [day, week, month, year]
            default: month
            example: month
        - name: startDate
          in: query
          description: Start date for custom period (ISO 8601)
          schema:
            type: string
            format: date-time
            example: 2023-01-01T00:00:00Z
        - name: endDate
          in: query
          description: End date for custom period (ISO 8601)
          schema:
            type: string
            format: date-time
            example: 2023-12-31T23:59:59Z
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          totalUsers:
                            type: integer
                            example: 1250
                          activeUsers:
                            type: integer
                            example: 1100
                          newUsers:
                            type: integer
                            example: 45
                          usersByRole:
                            type: object
                            properties:
                              USER:
                                type: integer
                                example: 1000
                              MODERATOR:
                                type: integer
                                example: 20
                              ADMIN:
                                type: integer
                                example: 5
                          usersByProvider:
                            type: object
                            properties:
                              LOCAL:
                                type: integer
                                example: 800
                              GOOGLE:
                                type: integer
                                example: 450
                          recentActivity:
                            type: array
                            items:
                              type: object
                              properties:
                                date:
                                  type: string
                                  example: 2023-01-15
                                count:
                                  type: integer
                                  example: 25
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden (Admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/bulk-update:
    patch:
      tags: [Users]
      summary: Bulk update users
      description: |
        Update multiple users at once with the same changes.

        **Features**:
        - Update role for multiple users
        - Activate/deactivate multiple users
        - Batch processing with error handling
        - Self-modification protection

        **Required Permission**: Admin only
      operationId: bulkUpdateUsers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userIds
                - updates
              properties:
                userIds:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 50
                  description: Array of user IDs to update
                  example: ["507f1f77bcf86cd799439011", "507f1f77bcf86cd799439012"]
                updates:
                  type: object
                  properties:
                    role:
                      type: string
                      enum: [USER, MODERATOR, ADMIN]
                      example: MODERATOR
                    isActive:
                      type: boolean
                      example: true
                  description: Updates to apply to all specified users
            examples:
              promote_to_moderator:
                summary: Promote users to moderator
                value:
                  userIds: ["507f1f77bcf86cd799439011", "507f1f77bcf86cd799439012"]
                  updates:
                    role: MODERATOR
              deactivate_users:
                summary: Deactivate multiple users
                value:
                  userIds: ["507f1f77bcf86cd799439013", "507f1f77bcf86cd799439014"]
                  updates:
                    isActive: false
      responses:
        '200':
          description: Bulk update completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          updated:
                            type: integer
                            description: Number of users successfully updated
                            example: 2
                          errors:
                            type: array
                            items:
                              type: string
                            description: List of errors encountered
                            example: []
        '400':
          description: Validation error or self-modification attempt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden (Admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user details by ID
      description: |
        Retrieve detailed information about a specific user by their ID.

        **Required Permission**: Moderator or Admin
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
              examples:
                success:
                  summary: User details
                  value:
                    success: true
                    message: User details retrieved successfully
                    data:
                      id: 507f1f77bcf86cd799439011
                      email: user@example.com
                      username: john_doe
                      firstName: John
                      lastName: Doe
                      role: USER
                      isActive: true
                      authProvider: LOCAL
                      isEmailVerified: true
                      createdAt: 2023-01-01T00:00:00Z
                      updatedAt: 2023-01-01T12:00:00Z
                      lastLoginAt: 2023-01-01T12:00:00Z
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden (Moderator/Admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    put:
      tags: [Users]
      summary: Update user by ID (Admin)
      description: |
        Update any user's information by their ID. Admin-only operation.

        **Features**:
        - Update user profile information
        - Change user role
        - Activate/deactivate user account
        - Self-modification protection

        **Required Permission**: Admin only
      operationId: updateUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID to update
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  maxLength: 50
                  example: John
                lastName:
                  type: string
                  maxLength: 50
                  example: Doe
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: '^[a-zA-Z0-9_]+$'
                  example: john_doe_admin
                profilePicture:
                  type: string
                  format: uri
                  example: https://example.com/avatar.jpg
                role:
                  type: string
                  enum: [USER, MODERATOR, ADMIN]
                  example: MODERATOR
                isActive:
                  type: boolean
                  example: true
            examples:
              promote_user:
                summary: Promote user to moderator
                value:
                  role: MODERATOR
              update_profile:
                summary: Update user profile
                value:
                  firstName: John
                  lastName: Doe
                  username: john_doe_updated
              deactivate_user:
                summary: Deactivate user
                value:
                  isActive: false
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: Validation error or self-modification attempt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden (Admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags: [Users]
      summary: Delete user by ID (Admin)
      description: |
        Permanently delete a user account by ID. Admin-only operation.

        **Warning**: This action cannot be undone.
        **Protection**: Cannot delete own account.

        **Required Permission**: Admin only
      operationId: deleteUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID to delete
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: User deleted
                  value:
                    success: true
                    message: User deleted successfully
        '400':
          description: Cannot delete own account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                self_deletion:
                  summary: Self-deletion attempt
                  value:
                    success: false
                    error:
                      message: Cannot delete your own account
                      statusCode: 400
                      status: error
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden (Admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/{id}/status:
    patch:
      tags: [Users]
      summary: Toggle user active status
      description: |
        Toggle a user's active status (activate/deactivate).

        **Features**:
        - Activate inactive users
        - Deactivate active users
        - Self-modification protection

        **Required Permission**: Admin only
      operationId: toggleUserStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID to toggle status
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: User status toggled successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
              examples:
                activated:
                  summary: User activated
                  value:
                    success: true
                    message: User activated successfully
                    data:
                      id: 507f1f77bcf86cd799439011
                      isActive: true
                deactivated:
                  summary: User deactivated
                  value:
                    success: true
                    message: User deactivated successfully
                    data:
                      id: 507f1f77bcf86cd799439011
                      isActive: false
        '400':
          description: Cannot change own status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden (Admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/{id}/role:
    patch:
      tags: [Users]
      summary: Update user role
      description: |
        Update a user's role (USER, MODERATOR, ADMIN).

        **Features**:
        - Promote/demote users
        - Role-based access control
        - Self-modification protection

        **Required Permission**: Admin only
      operationId: updateUserRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID to update role
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum: [USER, MODERATOR, ADMIN]
                  description: New role for the user
                  example: MODERATOR
            examples:
              promote_to_moderator:
                summary: Promote to moderator
                value:
                  role: MODERATOR
              promote_to_admin:
                summary: Promote to admin
                value:
                  role: ADMIN
              demote_to_user:
                summary: Demote to user
                value:
                  role: USER
      responses:
        '200':
          description: User role updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
              examples:
                success:
                  summary: Role updated
                  value:
                    success: true
                    message: User role updated successfully
                    data:
                      id: 507f1f77bcf86cd799439011
                      role: MODERATOR
        '400':
          description: Cannot change own role or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden (Admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # Database Management Endpoints
...